Consolidate fetch threads endpoints (#4309)
* refactor endpoints for fetching threads

* fix tests + update UI

* remove unused handlers

* remove unused handlers

* lint

---------
